actually made checkWIn function work, further improved readability
checkWin did not work previous because it was checking equivalence of
each player's spaces rather if a winning state was a subset of the
spaces.

for readability, moved some variables to the top and instead of having
seperate print statement for the cases of someone winning/not winning,
winner is initalized to "NO ONE" and only changes if checkWin returns a
winner.